+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './ButtonLink';
 export * from './Nav';
+export * from './SiteBanners';
